requirements.txt
requests>=2.32.3
numpy>=2.1.1
ddddocr>=1.5.5
pillow>=10.4.0
loguru>=0.7.2